sdfsdf

+++ src/Bank.java
@@ -29,7 +29,7 @@ public abstract interface Bank {
 	public abstract void deposit(Double amount); // has different effect for differiendr
 									// accounts
 
-	public abstract void withdraw(int amount); // some accounts have penalites
+	public abstract void withdraw(Double amount); // some accounts have penalites
 
 	public void assignCustomerID();
